requirements.txt
streamlit
biopython
openpyxl
XlsxWriter
pandas
numpy
databricks-sql-connector
